requirements.txt
pyyaml==6.0.1
jinja2==3.1.2 